Upgrade Expertise to API V2 (#88)
* Add paper_id to create_dataset

* Test get by paper id

* Add get by ID to service test client

* Adjust tests

* Replace iterget with get_note

* Adjust tests to use get_note

* Check API for get_paper_notes

* Check API for get_publications

* Check API for expertise_helper

* Add version keyword documentation

* Fix incorrectly merged test

* Skip results from dummy submission

* Prepare API v2 Mock Data

* Check API for get_submissions

* Remove duplicate from merge error

* Add dataset tests for V2

* Add celery scaffolding to test

* Add field version to POST request

* Set version as required field

* Use V2 client in celery

* Add generic API test

* Use V2 mock client in celery task

* Fix tests

* Replace general test with journal format

* Upgrade client for command line program

* Add default value to README

* Move version to optional field

* Exceptions on not supported versions

* Remove value field

* Get empty dict for missing key

* Pass in V1 client by default

* Allow V2 client to be passed in

* Update mock clients for celery task

* Modify mock clients

* Combine notes from author

* Handle different note formats for expertise

* Check invitation/ID from both APIs

* Update handling note schemas

* Fix API v2 Tests

* Remove unused code

* Change formatting

* Add BASEURL_V2 to server config

* Raise not found in mock

* Change submission handling

* Add V2 invitation test

* Add baseurl V2 to command line

* Add baseurl V2 to API

* Fix original attr bug

* Change fetching attributes

* Pass client v2 to create_dataset

* Simplify getattr call

* Change exception handling for single papers

* Remove old code

+++ tests/test_expertise_service.py
@@ -439,6 +439,7 @@ class TestExpertiseService():
         response = response.json['results']
         for item in response:
             submission_id, profile_id, score = item['submission'], item['user'], float(item['score'])
+            assert submission_id != 'dummy'
             assert len(submission_id) >= 1
             assert len(profile_id) >= 1
             assert profile_id.startswith('~')
@@ -529,8 +530,10 @@ class TestExpertiseService():
 
         # Clean up directory
         shutil.rmtree(f"./tests/jobs/")
-        os.remove('pytest.log')
-        os.remove('default.log')
+        if os.path.isfile('pytest.log'):
+            os.remove('pytest.log')
+        if os.path.isfile('default.log'):
+            os.remove('default.log')
 
 # def test_elmo_queue(openreview_context, celery_session_app, celery_session_worker):
 #     test_client = openreview_context['test_client']
